Simplify functionality and allow base backups via HTTP API
The functionality and configuration of the WAL-E image is now greatly
simplified. The WAL-E container no longer depends on functionality in
the TimescaleDB container; for instance, there is no longar a lock
file to synchronize container startup when, e.g., restoring.

Instead, the WAL-E container simply runs the HTTP API, with added
capability to trigger base backups. Restoring is delegated to
application developers for more flexibility. For instance, it is
possible to run this image as an init container in Kubernetes with a
custom script. Similarly, base backups should be triggered externally
via the HTTP API, for instance, using a Kubernetes CronJob.