Add Makefile for Docker builds
New Docker images can now be built with `make docker-image` and
images can be pushed to DockerHub with `make docker-push` (credentials
required).

This also changes the default image name to `timescale/wal-e` instead
of `timescale/timescaledb-wale`. This is partly to reflect that we no
longer depend on the TimescaleDB image and instead just package WAL-E
in Docker.